requirements.txt
aioredis==2.0.1
aioredis-fastapi==1.0.0
anyio==3.6.1
async-timeout==4.0.2
bcrypt==4.0.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.1
greenlet==1.1.3
h11==0.13.0
idna==3.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3

PyYAML~=6.0